Address the local Administrators group by its well-known SID.
Names of security objects built into Windows change depending on the
installation-time language (and cannot be modified without reinstalling the
system). Luckily, Puppet now understands Windows security identifiers and there
is a well-known SID corresponding to Administrators regardless of the language
used.

Ideally the name of the built-in administrator user should be replaced by a SID
too, especially given this can change even more frequently than the group
(there is a Group Policy setting for that). However, that one is more tricky
because its middle part changes between machines.